👌 Adjustments from code review
