+++ appcues/src/main/java/com/appcues/debugger/ui/theme/AppcuesDarkColors.kt
@@ -21,7 +21,7 @@ private val Green300 = Color(color = 0xFF5EBEBE)
 
 // Composing a color palette for dark mode are based on foundation figma doc.
 // Visit: https://www.figma.com/file/8Sozh9JnseLTF5oWjvGfIp/Foundations?type=design&node-id=5619-2635&mode=design&t=56CipGcX1bXwpUOW-0
-internal fun darkAppcuesTheme() = AppcuesThemeColors(
+internal fun appcuesDarkColors() = AppcuesThemeColors(
     background = Neutral600,
     backgroundSelected = Neutral700,
     error = Pink400,
